+++ src/OpenTelemetry/Metrics/Tags.cs
@@ -24,7 +24,7 @@ namespace OpenTelemetry.Metrics
         {
             this.KeyValuePairs = keyValuePairs;
 
-#if NETSTANDARD2_1 || NET6_0_OR_GREATER
+#if NET6_0_OR_GREATER
             HashCode hashCode = default;
             for (int i = 0; i < this.KeyValuePairs.Length; i++)
             {