+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Add support for exporting histograms aggregated using the
   [Base2 Exponential Bucket Histogram Aggregation](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/sdk.md#base2-exponential-bucket-histogram-aggregation).
-  ([#TBD](https://github.com/open-telemetry/opentelemetry-dotnet/pull/TBD))
+  ([#4337](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4337))
 
 ## 1.5.0-alpha.1