+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.5.0-alpha.2
+
+Released 2023-Mar-31
+
 * Add support for exporting histograms aggregated using the
   [Base2 Exponential Bucket Histogram Aggregation](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/sdk.md#base2-exponential-bucket-histogram-aggregation).
   ([#4337](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4337))